init wallet service
