+++ app/includes/messages-conversation.tpl
@@ -1,7 +1,7 @@
 <div class="messagesConversation">
-    <h2>{{newMessageService.messagesConversation[0].from}}
+    <h2>{{messageService.messagesConversation[0].from}}
         <div class="addressIdenticon med float" title="Address Indenticon"
-             blockie-address="{{newMessageService.messagesConversation[0].from}}"></div>
+             blockie-address="{{messageService.messagesConversation[0].from}}"></div>
     </h2>
     <hr />
 
@@ -9,7 +9,7 @@
     <ol class="list-group">
         <li
                 class="list-group-item"
-                ng-repeat="message in newMessageService.messagesConversation | orderBy: '+time' track by $index">
+                ng-repeat="message in messageService.messagesConversation | orderBy: '+time' track by $index">
             <div class="row">
                 {{message.time | date: 'yyyy-MM-dd HH:mm:ss'}}
             </div>
