+++ app/scripts/controllers/walletBalanceCtrl.js
@@ -4,13 +4,13 @@ var walletBalanceCtrl = function ($scope, $sce,
                                   backgroundNodeService,
                                   modalService,
                                   coldStakingService,
-                                  newMessageService) {
+                                  messageService) {
     $scope.ajaxReq = ajaxReq;
     $scope.erc20Abi = require('../abiDefinitions/erc20abi.json');
     $scope.DEXNS = require('../abiDefinitions/etcAbi.json')[5];
     $scope.DEXNSAddress = $scope.DEXNS.address;
 
-    $scope.newMessageService = newMessageService;
+    $scope.messageService = messageService;
 
     $scope.modalService = modalService;
     $scope.coldStakingService = coldStakingService;
